Add proper enums for some `DenseVectorProperty` fields
