+++ specification/_types/mapping/complex.ts
@@ -20,7 +20,6 @@
 import { TimeSeriesMetricType } from '@_types/mapping/TimeSeriesMetricType'
 import { double, integer } from '@_types/Numeric'
 import { CorePropertyBase, IndexOptions } from './core'
-import { DenseVectorIndexOptions } from './DenseVectorIndexOptions'
 import { PropertyBase } from './Property'
 
 export class FlattenedProperty extends PropertyBase {
@@ -47,15 +46,6 @@ export class ObjectProperty extends CorePropertyBase {
   enabled?: boolean
   subobjects?: boolean
   type?: 'object'
-}
-
-export class DenseVectorProperty extends PropertyBase {
-  type: 'dense_vector'
-  element_type?: string
-  dims?: integer
-  similarity?: string
-  index?: boolean
-  index_options?: DenseVectorIndexOptions
 }
 
 export class AggregateMetricDoubleProperty extends PropertyBase {
